chore: Update tests.yml to use Poetry for pre-commit and pytest commands
Update the tests.yml file to use Poetry for the pre-commit and pytest commands. This ensures that the project's dependencies are installed correctly and that the pre-commit hooks are run using Poetry. The changes include replacing "pre-commit install" with "poetry run pre-commit install" and "pytest --cov=pdbeccdutils" with "poetry run pytest --cov=pdbeccdutils".