Resolves bugs in writing cif file and parity similarity calculation (#28)
* corrected the reversal of bondtype from zero to dative

* use mol from sanitized result

* enable parsing a subsets of ccds from components cif file

* fix: fix error in writing cif file

This fixes the error in generating inchikey while inchi is missing

* style: formating

* chore: removed unnecessary None

* doc: add download badge and CLC feature

* chore: add citation file

* fix: write rdkit properties if it is present only

* fix: correct use of length of list in python

* fix: change bondtype from dative to zero for parity method

SMARTS with dative bondtype fails to find substructures, hence dative bond types are changed to zero

* test: add HEM for parity test

* chore: linting and formatting

* test: removed parity test using HEM

to check if the segmentation fault in githubworkflow is due to this

* fix: remove PDBe from unichem mapping sources

* chore: Update dependencies and package management

Update the project's dependencies and package management to use Poetry instead of pip. This includes installing Poetry and running `poetry install --with tests` to install the project dependencies. Remove the installation of `rdkit==2023.09.6` and `pre-commit`.

* chore: Update pre-commit hooks and dependencies

Update the pre-commit hooks to use the Ruff pre-commit hook repository and remove the black and flake8 hooks. Also, add the rST Formatter hook from the rstfmt repository.

* chore: Update tests.yml to use Poetry for pre-commit and pytest commands

Update the tests.yml file to use Poetry for the pre-commit and pytest commands. This ensures that the project's dependencies are installed correctly and that the pre-commit hooks are run using Poetry. The changes include replacing "pre-commit install" with "poetry run pre-commit install" and "pytest --cov=pdbeccdutils" with "poetry run pytest --cov=pdbeccdutils".

* Minor formatting

* chore: Install pre-commit package in tests.yml

Add the installation of the pre-commit package in the tests.yml file to ensure that pre-commit hooks are run during the testing process. This will help catch and fix any code style or formatting issues before committing the changes.

* chore: Update docs and publish pipelines to use poetry

* test: add HEM to parity test

* bump up version number

* 🔥 remove __init__ file

* 🔧 update documentation action to use poetry

* 🔧 update publish action to use poetry

* 🔧 update test action to use poetry

* 🎨 move details from setup to pyproject.toml file

* ♻️ use version information from pyproject.toml

* 🩹 add CCDC to unichem resources

* ✏️ fix typos

* ✏️ fix typos

* ♻️ refactor configs

* 🎨 use single function to get properties of rdkit objects

* 🎨 use rdkit_object_property function insted of get_componet_atom_id

* ✨ get name of clc from entities

* 🩹 import importlib.metadata

* 🎨 linting and formatting

* 🩹 removed "data" from the path

* 🔧 add poetry.toml file

* 🔧 update poetry.lock file

* 🔧 remove pre-commit from test workflow

* 🎨 linting and formatting

* 📝 replace github downloads with pypi downloads

* bump up version

* 📝 update changelog for release 0.8.6

* 🔧 create hook to generate poetry.lock file

* 📝 update readme with installation using poetry

* 🔧 update the rdkit version number

* 🔧 update poetry lock file

---------

Co-authored-by: roshan <roshan@ebi.ac.uk>
Co-authored-by: Sreenath Sasidharan Nair <sreenath@ebi.ac.uk>